+++ emscripten.py
@@ -189,14 +189,13 @@ def set_memory(static_bump):
   settings.HEAP_BASE = align_memory(stack_high)
 
 
-def report_missing_symbols(pre):
-  # the initial list of missing functions are that the user explicitly exported
-  # but were not implemented in compiled code
-  missing = set(settings.USER_EXPORTED_FUNCTIONS) - set(asmjs_mangle(e) for e in settings.WASM_EXPORTS)
-
-  for requested in sorted(missing):
-    if (f'function {requested}(') not in pre:
-      diagnostics.warning('undefined', f'undefined exported symbol: "{requested}"')
+def report_missing_symbols(js_library_funcs):
+  # Report any symbol that was explicitly exported but is present neither
+  # as a native function nor as a JS library function.
+  defined_symbols = set(asmjs_mangle(e) for e in settings.WASM_EXPORTS).union(js_library_funcs)
+  missing = set(settings.USER_EXPORTED_FUNCTIONS) - defined_symbols
+  for symbol in sorted(missing):
+    diagnostics.warning('undefined', f'undefined exported symbol: "{symbol}"')
 
   # Special hanlding for the `_main` symbol
 
@@ -312,10 +311,6 @@ def emscript(in_wasm, out_wasm, outfile_js, memfile, DEBUG):
     t = time.time()
 
   forwarded_json = json.loads(forwarded_data)
-  # For the wasm backend the implementedFunctions from compiler.js should
-  # always be empty. This only gets populated for __asm function when using
-  # the JS backend.
-  assert not forwarded_json['Functions']['implementedFunctions']
 
   pre, post = glue.split('// EMSCRIPTEN_END_FUNCS')
 
@@ -324,7 +319,7 @@ def emscript(in_wasm, out_wasm, outfile_js, memfile, DEBUG):
   if settings.ASYNCIFY:
     exports += ['asyncify_start_unwind', 'asyncify_stop_unwind', 'asyncify_start_rewind', 'asyncify_stop_rewind']
 
-  report_missing_symbols(pre)
+  report_missing_symbols(forwarded_json['libraryFunctions'])
 
   if not outfile_js:
     logger.debug('emscript: skipping remaining js glue generation')
@@ -337,11 +332,8 @@ def emscript(in_wasm, out_wasm, outfile_js, memfile, DEBUG):
     # In regular runtime, atinits etc. exist in the preamble part
     pre = apply_static_code_hooks(forwarded_json, pre)
 
-  # merge forwarded data
-  settings.EXPORTED_FUNCTIONS = forwarded_json['EXPORTED_FUNCTIONS']
-
   asm_consts = create_asm_consts(metadata)
-  em_js_funcs = create_em_js(forwarded_json, metadata)
+  em_js_funcs = create_em_js(metadata)
   asm_const_pairs = ['%s: %s' % (key, value) for key, value in asm_consts]
   asm_const_map = 'var ASM_CONSTS = {\n  ' + ',  \n '.join(asm_const_pairs) + '\n};\n'
   pre = pre.replace(
@@ -476,7 +468,7 @@ def create_asm_consts(metadata):
   return asm_consts
 
 
-def create_em_js(forwarded_json, metadata):
+def create_em_js(metadata):
   em_js_funcs = []
   separator = '<::>'
   for name, raw in metadata.get('emJsFuncs', {}).items():
@@ -490,7 +482,6 @@ def create_em_js(forwarded_json, metadata):
     arg_names = [arg.split()[-1].replace("*", "") for arg in args if arg]
     func = 'function {}({}){}'.format(name, ','.join(arg_names), body)
     em_js_funcs.append(func)
-    forwarded_json['Functions']['libraryFunctions'][name] = 1
 
   return em_js_funcs
 
@@ -798,6 +789,7 @@ def load_metadata_wasm(metadata_raw, DEBUG):
   unexpected_exports = [asmjs_mangle(e) for e in unexpected_exports]
   unexpected_exports = [e for e in unexpected_exports if e not in settings.EXPORTED_FUNCTIONS]
   building.user_requested_exports.update(unexpected_exports)
+  settings.EXPORTED_FUNCTIONS.extend(unexpected_exports)
 
   return metadata